Task fairness schema (#7627)
## What changed?
Added new schema for task table that supports fairness scheduling.

* split `matching_task_store.go` into v1 (pre-fairness) and v2
(fairness) - clean copy seemed safest
* inject two task stores, `PhysicalTaskQueueManager` picks the one it
needs
