+++ common/persistence/client/fx.go
@@ -61,6 +61,7 @@ var Module = fx.Options(
 	fx.Provide(managerProvider(Factory.NewClusterMetadataManager)),
 	fx.Provide(managerProvider(Factory.NewMetadataManager)),
 	fx.Provide(managerProvider(Factory.NewTaskManager)),
+	fx.Provide(managerProvider(Factory.NewFairTaskManager)),
 	fx.Provide(managerProvider(Factory.NewNamespaceReplicationQueue)),
 	fx.Provide(managerProvider(Factory.NewShardManager)),
 	fx.Provide(managerProvider(Factory.NewExecutionManager)),
@@ -161,7 +162,6 @@ func DataStoreFactoryProvider(
 	metricsHandler metrics.Handler,
 	tracerProvider trace.TracerProvider,
 ) persistence.DataStoreFactory {
-
 	var dataStoreFactory persistence.DataStoreFactory
 	defaultStoreCfg := cfg.DataStores[cfg.DefaultStore]
 	switch {
