+++ common/persistence/data_interfaces.go
@@ -1166,6 +1166,7 @@ type (
 		GetTaskQueuesByBuildId(ctx context.Context, request *GetTaskQueuesByBuildIdRequest) ([]string, error)
 		CountTaskQueuesByBuildId(ctx context.Context, request *CountTaskQueuesByBuildIdRequest) (int, error)
 	}
+	FairTaskManager TaskManager
 
 	// MetadataManager is used to manage metadata CRUD for namespace entities
 	MetadataManager interface {
