+++ common/persistence/persistence_interface.go
@@ -33,6 +33,8 @@ type (
 		Close()
 		// NewTaskStore returns a new task store
 		NewTaskStore() (TaskStore, error)
+		// NewTaskStore returns a new task store
+		NewTaskFairnessStore() (TaskStore, error)
 		// NewShardStore returns a new shard store
 		NewShardStore() (ShardStore, error)
 		// NewMetadataStore returns a new metadata store
@@ -61,20 +63,24 @@ type (
 	TaskStore interface {
 		Closeable
 		GetName() string
+
 		CreateTaskQueue(ctx context.Context, request *InternalCreateTaskQueueRequest) error
 		GetTaskQueue(ctx context.Context, request *InternalGetTaskQueueRequest) (*InternalGetTaskQueueResponse, error)
 		UpdateTaskQueue(ctx context.Context, request *InternalUpdateTaskQueueRequest) (*UpdateTaskQueueResponse, error)
 		ListTaskQueue(ctx context.Context, request *ListTaskQueueRequest) (*InternalListTaskQueueResponse, error)
 		DeleteTaskQueue(ctx context.Context, request *DeleteTaskQueueRequest) error
+		GetTaskQueuesByBuildId(ctx context.Context, request *GetTaskQueuesByBuildIdRequest) ([]string, error)
+		CountTaskQueuesByBuildId(ctx context.Context, request *CountTaskQueuesByBuildIdRequest) (int, error)
+
 		CreateTasks(ctx context.Context, request *InternalCreateTasksRequest) (*CreateTasksResponse, error)
 		GetTasks(ctx context.Context, request *GetTasksRequest) (*InternalGetTasksResponse, error)
 		CompleteTasksLessThan(ctx context.Context, request *CompleteTasksLessThanRequest) (int, error)
+
 		GetTaskQueueUserData(ctx context.Context, request *GetTaskQueueUserDataRequest) (*InternalGetTaskQueueUserDataResponse, error)
 		UpdateTaskQueueUserData(ctx context.Context, request *InternalUpdateTaskQueueUserDataRequest) error
 		ListTaskQueueUserDataEntries(ctx context.Context, request *ListTaskQueueUserDataEntriesRequest) (*InternalListTaskQueueUserDataEntriesResponse, error)
-		GetTaskQueuesByBuildId(ctx context.Context, request *GetTaskQueuesByBuildIdRequest) ([]string, error)
-		CountTaskQueuesByBuildId(ctx context.Context, request *CountTaskQueuesByBuildIdRequest) (int, error)
 	}
+
 	// MetadataStore is a lower level of MetadataManager
 	MetadataStore interface {
 		Closeable
@@ -300,6 +306,7 @@ type (
 
 	InternalCreateTask struct {
 		TaskId     int64
+		Pass       int64
 		ExpiryTime *timestamppb.Timestamp
 		Task       *commonpb.DataBlob
 		Subqueue   int
