+++ common/persistence/sql/factory.go
@@ -73,6 +73,16 @@ func (f *Factory) NewTaskStore() (p.TaskStore, error) {
 	return newTaskPersistence(conn, f.cfg.TaskScanPartitions, f.logger)
 }
 
+// NewTaskFairnessStore returns a new task store
+// TODO(fairness): cleanup; rename to NewTaskStore
+func (f *Factory) NewTaskFairnessStore() (p.TaskStore, error) {
+	conn, err := f.mainDBConn.Get()
+	if err != nil {
+		return nil, err
+	}
+	return newTaskPersistence(conn, f.cfg.TaskScanPartitions, f.logger)
+}
+
 // NewShardStore returns a new shard store
 func (f *Factory) NewShardStore() (p.ShardStore, error) {
 	conn, err := f.mainDBConn.Get()
