+++ common/persistence/tests/cassandra_test.go
@@ -212,6 +212,20 @@ func TestCassandraTaskQueueSuite(t *testing.T) {
 	suite.Run(t, s)
 }
 
+// TODO(fairness): cleanup; rename to TestCassandraTaskQueueSuite
+func TestCassandraTaskFairnessQueueSuite(t *testing.T) {
+	testData, tearDown := setUpCassandraTest(t)
+	defer tearDown()
+
+	taskQueueStore, err := testData.Factory.NewTaskFairnessStore()
+	if err != nil {
+		t.Fatalf("unable to create Cassandra DB: %v", err)
+	}
+
+	s := NewTaskQueueSuite(t, taskQueueStore, testData.Logger)
+	suite.Run(t, s)
+}
+
 func TestCassandraTaskQueueTaskSuite(t *testing.T) {
 	testData, tearDown := setUpCassandraTest(t)
 	defer tearDown()
