+++ schema/cassandra/temporal/schema.cql
@@ -80,6 +80,7 @@ CREATE TABLE history_tree (
 };
 
 -- Stores activity or workflow tasks
+-- TODO(fairness): cleanup; remove this table
 CREATE TABLE tasks (
   namespace_id        uuid,
   task_queue_name     text,
@@ -95,6 +96,26 @@ CREATE TABLE tasks (
 ) WITH COMPACTION = {
     'class': 'org.apache.cassandra.db.compaction.LeveledCompactionStrategy'
   };
+
+-- Stores activity or workflow tasks
+-- Used for fairness scheduling. (pass, task_id) are monotonically increasing.
+CREATE TABLE tasks_v2 (
+  namespace_id        uuid,
+  task_queue_name     text,
+  task_queue_type     int, -- enum TaskQueueType {ActivityTask, WorkflowTask}
+  type                int, -- enum rowType {Task, TaskQueue}
+  pass                bigint, -- pass for tasks (see stride scheduling algorithm for fairness)
+  task_id             bigint, -- unique identifier for tasks
+  range_id            bigint, -- used to ensure that only one process can write to the table
+  ack_level           bigint, -- ack level for the task queue
+  task                blob,
+  task_encoding       text,
+  task_queue          blob,
+  task_queue_encoding text,
+PRIMARY KEY ((namespace_id, task_queue_name, task_queue_type), type, pass, task_id)
+) WITH COMPACTION = {
+    'class': 'org.apache.cassandra.db.compaction.LeveledCompactionStrategy'
+};
 
 -- Stores task queue information such as user provided versioning data
 -- OR
