+++ service/matching/handler.go
@@ -52,6 +52,7 @@ type (
 		Logger                        log.Logger
 		ThrottledLogger               log.Logger
 		TaskManager                   persistence.TaskManager
+		FairTaskManager               persistence.FairTaskManager
 		HistoryClient                 resource.HistoryClient
 		MatchingRawClient             resource.MatchingRawClient
 		DeploymentStoreClient         deployment.DeploymentStoreClient
@@ -91,6 +92,7 @@ func NewHandler(
 		throttledLogger: params.ThrottledLogger,
 		engine: NewEngine(
 			params.TaskManager,
+			params.FairTaskManager,
 			params.HistoryClient,
 			params.MatchingRawClient, // Use non retry client inside matching
 			params.DeploymentStoreClient,
